Add timeline command
After this I'll make it possible to call single user timeline, like
`pyxis timeline <user>'. We will check in early in code & read only
this file.

Also, need to append user information in `%twtxt' along with support
for multiple entries in single date. Currently only one entry will
show up if there is a date clash.

I'll fix these by changing the data structure to hash table of a list
of hash table. So the list will contain 2 hashes - user & entry & the
list of this will be store in hash table so as to not over write
multiple entries of same time.